requirements.txt
shortuuid
redis
rq
Flask
flask-cors
pillow